board select max 수정_최종

+++ src/main/java/com/green/boardver3/board/BoardService.java
@@ -28,7 +28,7 @@ public class BoardService {
     }
 
     public int selLastBoard(int row) {
-        int count = mapper.selLastBoard(row);
+        double count = mapper.selLastBoard(row);
         return (int)Math.ceil(count/ row);
     }
 }
